Add a 'Tgt' button to the Cruise Waypoint Controller.
If a target (located within the same spehere of influence) has been
selected, this button captures the lat/lon of the target and sets those as
the current waypoint, updating the lat/lon display.  If the target is not
'Landed', a warning is issued but the current coordinates of the target are
captured for use, anyway.

This change also permits the input of degree-minute-second format lat/lon
coordinates directly into the corresponding input boxes.  A leading polarity
plus/minus sign disqualifies the input as d-m-s format.  Decimal format is
permitted and assumed until a trailing space or direction indicator is
typed.  No real-time update of coordinates occurs until the trailing
indicator -- 'N' or 'S' for latitude, 'E' or 'W' for longitude -- is typed.

The syntax is deg [min [sec]]{N|S|E|W}, e.g. 34 17 22S

When correct d-m-s formatted input is entered, after a pause, the input box
is updated to the decimal equivalent to 4 decimal places.